requirements.txt
httplib2
oauth2client
google-api-python-client
py-dotenv==1.0.0
bs4==0.0.1
openai==1.2.4

